+++ metadata.rb
@@ -6,4 +6,4 @@ maintainer_email  'nuspl@nvwls.com'
 license           'Apache 2.0'
 description       'Provides HWRPs for creating authoritative resources'
 long_description  IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version           '0.7.0'
+version           '0.8.0'