+++ spec/cookbooks/test/attributes/default.rb
@@ -1,0 +1,4 @@
+# encoding: UTF-8
+
+default['directory']['recursive'] = nil
+default['directory']['pattern'] = nil