+++ spec/cookbooks/test/metadata.rb
@@ -1,0 +1,6 @@
+# encoding: UTF-8
+
+name		'test'
+version		'0.0.1'
+
+depends		'zap', '>= 0.7.0'